+++ app/src/Playground/index.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable no-console */
 
-import { useState } from "react";
-import { useForm, useControlled } from "react-cool-form";
+import { useForm, useFieldArray, useControlled } from "react-cool-form";
 
 const Field = ({ name, ...rest }: any) => {
   const [props] = useControlled(name, rest);
@@ -9,24 +8,32 @@ const Field = ({ name, ...rest }: any) => {
 };
 
 export default () => {
-  const [show, setShow] = useState(false);
   const { form, reset } = useForm({
-    defaultValues: { foo: ["1", "2"] },
+    defaultValues: {
+      foo: [
+        { a: "a1", b: "b1" },
+        { a: "a2", b: "b2" },
+      ],
+    },
     onSubmit: (values) => console.log("onSubmit: ", values),
   });
+  const [fields, { remove }] = useFieldArray("foo");
+
+  console.log("LOG ===> Re-render");
 
   return (
     <form ref={form}>
-      <input name="foo" type="checkbox" value="1" defaultChecked />
-      {show && <input name="foo" type="checkbox" value="2" defaultChecked />}
+      {fields.map((fieldName) => (
+        <div key={fieldName}>
+          <input name={`${fieldName}.a`} />
+          <Field name={`${fieldName}.b`} />
+        </div>
+      ))}
       <input type="submit" />
-      <button type="button" onClick={() => setShow(!show)}>
-        Toggle
+      <button type="button" onClick={() => remove(0)}>
+        Remove
       </button>
-      <button
-        type="button"
-        onClick={() => console.log("LOG ===> ", reset({ foo: [] }))}
-      >
+      <button type="button" onClick={() => reset()}>
         Reset
       </button>
     </form>