+++ src/hooks/useState.ts
@@ -6,9 +6,7 @@ import {
   FormState,
   FormStateReturn,
   Observer,
-  ObserverHandler,
   SetStateRef,
-  SetUsedState,
 } from "../types";
 import useLatest from "./useLatest";
 import { get, getIsDirty, isEmptyObject, set } from "../utils";
@@ -19,13 +17,13 @@ export default <V>(
 ): FormStateReturn<V> => {
   const [, forceUpdate] = useReducer((c) => c + 1, 0);
   const stateRef = useRef(initialState);
-  const stateObserversRef = useRef<Observer<V>[]>([
+  const observersRef = useRef<Observer<V>[]>([
     { usedState: {}, notify: forceUpdate },
   ]);
   const onChangeRef = useLatest(onChange || (() => undefined));
 
   const setStateRef = useCallback<SetStateRef>(
-    (path, value, { fieldPath, shouldUpdate = true } = {}) => {
+    (path, value, { fieldPath, shouldSkipUpdate, shouldForceUpdate } = {}) => {
       const key = path.split(".")[0];
 
       if (!key) {
@@ -33,10 +31,10 @@ export default <V>(
           stateRef.current = value;
           onChangeRef.current(stateRef.current);
 
-          stateObserversRef.current.forEach(({ usedState, notify }) => {
-            if (shouldUpdate && !isEmptyObject(usedState))
-              notify(stateRef.current);
-          });
+          observersRef.current.forEach(
+            ({ usedState, notify }) =>
+              !isEmptyObject(usedState) && notify(stateRef.current)
+          );
         }
 
         return;
@@ -52,56 +50,37 @@ export default <V>(
           dirty,
           isDirty: prevIsDirty,
           isValid: prevIsValid,
+          submitCount: prevSubmitCount,
         } = state;
-        let { submitCount: prevSubmitCount } = state;
+        let { submitCount: currSubmitCount } = state;
         const isDirty = key === "dirty" ? getIsDirty(dirty) : prevIsDirty;
         const isValid = key === "errors" ? isEmptyObject(errors) : prevIsValid;
         const submitCount =
           key === "isSubmitting" && value
-            ? (prevSubmitCount += 1)
-            : prevSubmitCount;
+            ? (currSubmitCount += 1)
+            : currSubmitCount;
 
         stateRef.current = { ...state, isDirty, isValid, submitCount };
         onChangeRef.current(stateRef.current);
 
-        if (!shouldUpdate) return;
+        if (shouldSkipUpdate) return;
 
         path = fieldPath || path;
-        stateObserversRef.current.forEach(({ usedState, notify }) => {
-          if (
-            Object.keys(usedState).some(
-              (k) => path.startsWith(k) || k.startsWith(path)
-            ) ||
-            (usedState.isDirty && isDirty !== prevIsDirty) ||
-            (usedState.isValid && isValid !== prevIsValid)
-          )
-            notify(stateRef.current);
-        });
+        observersRef.current.forEach(
+          ({ usedState, notify }) =>
+            (shouldForceUpdate ||
+              Object.keys(usedState).some(
+                (k) => path.startsWith(k) || k.startsWith(path)
+              ) ||
+              (usedState.isDirty && isDirty !== prevIsDirty) ||
+              (usedState.isValid && isValid !== prevIsValid) ||
+              (usedState.submitCount && submitCount !== prevSubmitCount)) &&
+            notify(stateRef.current)
+        );
       }
     },
     [onChangeRef]
   );
 
-  const setUsedState = useCallback<SetUsedState>((usedState) => {
-    stateObserversRef.current[0].usedState = usedState;
-  }, []);
-
-  const subscribeObserver = useCallback<ObserverHandler<V>>(
-    (observer) => stateObserversRef.current.push(observer),
-    []
-  );
-
-  const unsubscribeObserver = useCallback<ObserverHandler<V>>((observer) => {
-    stateObserversRef.current = stateObserversRef.current.filter(
-      (o) => o !== observer
-    );
-  }, []);
-
-  return {
-    stateRef,
-    setStateRef,
-    setUsedState,
-    subscribeObserver,
-    unsubscribeObserver,
-  };
+  return { stateRef, setStateRef, observersRef };
 };